requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
selenium>=4.0.0
python-dotenv>=0.19.0
translate>=3.6.1
pytz>=2021.3
webdriver-manager>=3.8.0
pymongo>=4.0.0 